MQTT connection created in native from JNI
Still needs to be returned, that's a problem for future me

+++ java/software/amazon/awssdk/crt/mqtt/MqttConnection.java
@@ -34,7 +34,7 @@ public final class MqttConnection implements AutoCloseable {
         public String clientId = "";
         public boolean cleanSession = true;
         public short keepAliveMs = 0;
-        public long timeout = 1000; 
+        public short timeout = 1000; 
 
         public ConnectOptions() {
         }
